% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{abs.interval}
\alias{abs.interval}
\title{Absolute value}
\usage{
\method{abs}{interval}(x)
}
\arguments{
\item{x}{An interval}
}
\value{
The numeric result as interval or number.
}
\description{
Absolute value
}
\examples{
abs(...(4, 9))
abs(-4 \%...\% 9)
abs(interval(4, 9))

}
