% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pbinom_intarith}
\alias{pbinom_intarith}
\title{The Binomial Distribution}
\usage{
pbinom_intarith(x, size, prob)
}
\arguments{
\item{x}{Number of successes}

\item{size}{Number of trials}

\item{prob}{Probability of success on each trial}
}
\value{
The numeric result as interval or number.
}
\description{
For more information, refer to the documentation
of \code{\link[=dbinom]{dbinom()}}, \code{\link[=pbinom]{pbinom()}}, \code{\link[=qbinom]{qbinom()}}.
}
\examples{
pbinom_intarith(...(5, 6), ...(20, 21), ...(0.3, 0.5))
}
