% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{...}
\alias{...}
\title{Dots symbol constructor for intervals}
\usage{
...(l, u)
}
\arguments{
\item{l}{The lower bound of the interval.}

\item{u}{The upper bound of the interval.}
}
\value{
An object of class interval.
}
\description{
One of the constructor functions for objects of class interval.
}
\examples{
...(1, 2)

}
