% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{-.interval}
\alias{-.interval}
\title{Subtraction}
\usage{
\method{-}{interval}(x, y = NULL)
}
\arguments{
\item{x}{Left operand}

\item{y}{Right operand}
}
\value{
The numeric result as interval or number.
}
\description{
If two arguments are provided, performs subtraction.
#' If only one is provided, acts as unary minus.
}
\examples{
...(3, 4) - ...(1, 2)
1 \%...\% 2 - 3 \%...\% 4
interval(1, 2) - interval(3, 4)
-...(3, 4)
-(1 \%...\% 2)
-interval(1, 2)

}
