% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{exp.interval}
\alias{exp.interval}
\title{Exponential}
\usage{
\method{exp}{interval}(x)
}
\arguments{
\item{x}{Exponent}
}
\value{
The numeric result as interval or number.
}
\description{
Exponential
}
\examples{
exp(...(3, 4))
exp(3 \%...\% 4)
exp(interval(3, 4))

}
