% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intarith.R
\name{intarith}
\alias{intarith}
\title{Evaluate expression}
\usage{
intarith(expr, flags = NULL, env = globalenv())
}
\arguments{
\item{expr}{An arithmetic expression as string.}

\item{flags}{Optional flags.}

\item{env}{The environment to be used for the evaluation.}
}
\value{
The numeric result of the expression as interval.
}
\description{
Evaluate an expression in interval arithmetic without having to create
interval objects. Intervals can be written as "...(a, b)" or "a...b".
}
\examples{
intarith("1...2 + ...(3, 4) + 5")

}
