\name{interval-package}
\alias{interval}
\docType{package}
\title{
Tests and NPMLE for interval censored data
}
\description{
The main functions are \code{\link{icfit}} to fit nonparametric survival curves together 
with \code{\link{plot.icfit}} to 
plot them, and \code{\link{ictest}} to perform logrank or Wilcoxon type tests.
}
\details{
\tabular{ll}{
Package: \tab interval\cr
Type: \tab Package\cr
Version: \tab 1.1-0.8\cr
Date: \tab 2021-10-08\cr
License: \tab GPL 2\cr
}
}
\author{
Michael Fay

Maintainer: M.P. Fay <mfay@niaid.nih.gov>
}
\references{
    Fay, MP and Shaw, PA (2010). Exact and Asymptotic Weighted Logrank Tests for Interval Censored Data: The interval R package.
    Journal of Statistical Software. \doi{10.18637/jss.v036.i02}. 36 (2):1-34.

}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\examples{}
