% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{interval}
\alias{interval}
\title{Interval}
\usage{
interval(expr, flags = NULL, env = globalenv())
}
\arguments{
\item{expr}{An arithmetic expression.}
}
\value{
The numeric result of the expression.
}
\description{
Interval
}
