% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Normal}
\alias{Normal}
\alias{pnorm}
\alias{qnorm}
\alias{dnorm}
\title{The Normal Distribution}
\usage{
pnorm(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

qnorm(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

dnorm(x, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{q, x}{vector of quantiles.}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\emph{P}[\emph{X≤x}], otherwise, \emph{P}[\emph{X≥x}]}

\item{log.p, log}{logical; if TRUE, probabilities p are given as log(p)}

\item{p}{vector of probabilities}
}
\value{
The numeric result as interval or number.
}
\description{
For more information, refer to the documentation
of \code{\link[=dnorm]{dnorm()}}, \code{\link[=pnorm]{pnorm()}}, \code{\link[=qnorm]{qnorm()}} from \code{stats}.
}
\examples{
pnorm(...(8, 9), ...(10, 11), ...(2, 3))
qnorm(...(0.3, 0.4), ...(10, 11), ...(2, 3))
dnorm(...(8, 9), ...(10, 11), ...(2, 3))
}
