% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{^.interval}
\alias{^.interval}
\title{Power}
\usage{
\method{^}{interval}(x, y)
}
\arguments{
\item{x}{Base}

\item{y}{Exponent as number}
}
\value{
The numeric result as interval or number.
}
\description{
Power
}
\examples{
...(3, 4) ^ 2
3 \%...\% 4 ^ 3
interval(3, 4) ^ 2

}
