% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.interval}
\alias{print.interval}
\title{Print intervals}
\usage{
\method{print}{interval}(x, ...)
}
\arguments{
\item{x}{An object of class interval.}

\item{...}{Further arguments.}
}
\description{
Print intervals as "lower...upper".
}
