% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round.interval}
\alias{round.interval}
\title{Round}
\usage{
\method{round}{interval}(x, digits = getOption("digits"))
}
\arguments{
\item{x}{An interval}

\item{digits}{The number of digits to keep after the decimal comma}
}
\value{
The numeric result as interval.
}
\description{
Round
}
\examples{
round(...(4, 9))
round(-4 \%...\% 9)
round(interval(4, 9))

}
