% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round2interval}
\alias{round2interval}
\title{Round to interval}
\usage{
round2interval(x, digits = getOption("digits"))
}
\arguments{
\item{x}{A number}

\item{digits}{The number of digits to keep after the decimal comma}
}
\value{
The numeric result as interval.
}
\description{
The number is converted to a degenerated interval before rounding.
This ensures that the result is still an interval.
}
\examples{
round2interval(4.166, 2)

}
