% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{sin.interval}
\alias{sin.interval}
\title{Sin}
\usage{
\method{sin}{interval}(x)
}
\arguments{
\item{x}{An interval}
}
\value{
The numeric result as interval or number.
}
\description{
Sin
}
\examples{
sin(...(4, 9))
sin(-4 \%...\% 9)
sin(interval(4, 9))

}
