% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{/.interval}
\alias{/.interval}
\title{Division}
\usage{
\method{/}{interval}(x, y)
}
\arguments{
\item{x}{Left operand}

\item{y}{Right operand}
}
\value{
The numeric result as interval or number.
}
\description{
Division
}
\examples{
...(3, 4) / ...(1, 2)
1 \%...\% 2 / 3 \%...\% 4
interval(1, 2) / interval(3, 4)

}
