% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{sqrt.interval}
\alias{sqrt.interval}
\title{Square root (standard)}
\usage{
\method{sqrt}{interval}(x)
}
\arguments{
\item{x}{Radicand}
}
\value{
The numeric result as interval or number.
}
\description{
Negative bounds return NaN.
}
\examples{
sqrt(...(4, 9))
sqrt(4 \%...\% 9)
sqrt(interval(4, 9))

}
