% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-ops.R
\name{sqrt0}
\alias{sqrt0}
\title{Square root (with cropping)}
\usage{
sqrt0(x)
}
\arguments{
\item{x}{Radicand}
}
\value{
The numeric result as interval or number.
}
\description{
Lower bound of the radicand is set to 0 in case it is negative.
Otherwise it is equivalent to the \code{sqrt} function.
}
\examples{
sqrt0(...(4, 9))
sqrt0(-4 \%...\% 9)
sqrt0(interval(4, 9))

}
