% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{TDist}
\alias{TDist}
\alias{pt}
\title{The Student t Distribution}
\usage{
pt(q, df, ncp = 0, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{df}{degree of freedoms}

\item{ncp}{non-centrality parameter. This is only provided for the base \code{stats::pt} function.
Not supported by the interval version of \code{pt}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\emph{P}[\emph{X≤x}], otherwise, \emph{P}[\emph{X≥x}]}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{p}{vector of probabilities}
}
\value{
The numeric result as interval or number.
}
\description{
For more information, refer to the documentation
of \code{\link[=dt]{dt()}}, \code{\link[=pt]{pt()}}, \code{\link[=qt]{qt()}} from \code{stats}.
}
\examples{
pt(...(2, 3), ...(10, 11))
}
