% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Binomial}
\alias{Binomial}
\alias{pbinom}
\alias{qbinom}
\alias{dbinom}
\title{The Binomial Distribution}
\usage{
pbinom(x, size, prob, lower.tail = TRUE, log.p = FALSE)

qbinom(p, size, prob, lower.tail = TRUE, log.p = FALSE)

dbinom(x, size, prob, ...)
}
\arguments{
\item{x}{Number of successes}

\item{size}{Number of trials}

\item{prob}{Probability of success on each trial}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\emph{P}[\emph{X≤x}], otherwise, \emph{P}[\emph{X≥x}]}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{p}{probability}

\item{...}{further arguments}
}
\value{
The numeric result as interval or number.
}
\description{
For more information, refer to the documentation
of \code{\link[=dbinom]{dbinom()}}, \code{\link[=pbinom]{pbinom()}}, \code{\link[=qbinom]{qbinom()}} from \code{stats}.
}
\examples{
dbinom(...(5, 6), ...(20, 21), ...(0.3, 0.5))
pbinom(...(5, 6), ...(20, 21), ...(0.3, 0.5))
qbinom(...(0.5, 0.6), ...(20, 21), ...(0.3, 0.5))
}
