\name{interval-package}
\alias{interval}
\docType{package}
\title{
Tests and NPMLE for interval censored data
}
\description{
The main functions are \code{\link{icfit}} to fit nonparametric survival curves together 
with \code{\link{plot.icfit}} to 
plot them, and \code{\link{ictest}} to perform logrank or Wilcoxon type tests.
}
\details{
\tabular{ll}{
Package: \tab interval\cr
Type: \tab Package\cr
Version: \tab 0.9-9.0\cr
Date: \tab 2009-11-06\cr
License: \tab GPL 2\cr
}
}
\author{
Michael Fay

Maintainer: M.P. Fay <mfay@niaid.nih.gov>
}
%\references{~~ Literature or other references for background information ~~}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\examples{}
