% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{interval}
\alias{interval}
\title{Constructor for intervals}
\usage{
interval(l, u)
}
\arguments{
\item{l}{The lower bound of the interval.}

\item{u}{The upper bound of the interval.}
}
\value{
An object of class interval.
}
\description{
Constructor function for objects of class interval.
}
\examples{
interval(1,2)

}
