% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intarith.R
\name{interval}
\alias{interval}
\title{Interval}
\usage{
interval(expr, flags = NULL, env = globalenv())
}
\arguments{
\item{expr}{An arithmetic expression.}

\item{flags}{Optional flags.}

\item{env}{The environment to be used for the evaluation.}
}
\value{
The numeric result of the expression.
}
\description{
Interval
}
